changed _ptmouse(short to float), made changes to pathFinding

+++ Classes/Direct2DFrameWork.cpp
@@ -109,8 +109,8 @@ LRESULT WndProc(HWND hWnd, UINT iMessage, WPARAM wParam, LPARAM lParam)
 		GetWindowRect(_hWnd,&windowRc);
 		RECT windowClnt;
 		GetClientRect(_hWnd, &windowClnt);
-		_ptMouse.x = (short)LOWORD(lParam) * (windowRc.right - windowRc.left) / (windowClnt.right - windowClnt.left);
-		_ptMouse.y = (short)HIWORD(lParam) * (windowRc.bottom - windowRc.top) / (windowClnt.bottom - windowClnt.top);
+		_ptMouse.x = (float)LOWORD(lParam) * (windowRc.right - windowRc.left) / (windowClnt.right - windowClnt.left);
+		_ptMouse.y = (float)HIWORD(lParam) * (windowRc.bottom - windowRc.top) / (windowClnt.bottom - windowClnt.top);
 
 		break;
 	case WM_KEYDOWN:
